set ensureThatRecordExists to false for surveyLink
Apparently, my changes in #29 caused an issue. `REDCap::getSurveyLink()`, pre-version 10.6.1 (lowest I could find), checks if the record exists in a specific arm and will not proceed. In newer versions of REDCap, `REDCap::getSurveyLink()` introduced `$ensureThatRecordExists` parameter that skips that check. This PR checks the version of REDCap in use and sets that parameter to false. 

Otherwise, passthrough links will fail if the record doesn't already exist in that arm. 

The previous implementation of the surveyPassthru functionality did not have this problem. There is no fix for versions before 10.6.1 at the moment.
 
I'm not sure of the best way of handling it for older versions; maybe raise the minimum version for the EM?